thesis-checker-web/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-docx==1.1.0
PyPDF2==3.0.1
markdown==3.5.2
python-multipart==0.0.6
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
motor==3.3.2
pymongo==4.6.1
